Better structure by putting types in more appropriate files and reexporting
Follows up #5 to keep its API, but undo a lot of the file structure
changes. A big downside of #5 is that a lot o the core code ends up in
`__init__.py` files, which are really hard to fast find because they all
have the same name, and generally just awkward to work with.

Python docs make this impossible to find, but I found a trick in
SQLAlchemy where it's possible to reexport types from a module using an
`as` in an `import` like this:

    # Reexport for more ergonomic use in calling code.
    from .driver_protocol import (
        GetParams as GetParams,
        JobInsertParams as JobInsertParams,
        DriverProtocol as DriverProtocol,
    )

Here, move code back into files with names better suited for it (e.g.
`Client` goes back in `client.py`) and use this trick to reexport
user-facing types so that we keep the same import ergonomic benefits as
laid out in #5.