Enable job insertion on both transactions and top level pool
A problem with the API currently is that is that inserts on a top-level
SQLAlchemy engine aren't supported. In the test suite, the driver is
always initialized with a transaction:

    with engine.begin() as tx:
        yield riversqlalchemy.Driver(tx)

But if trying to initialize the driver with the engine itself, we'd see
failures because it doesn't support the necessary SQL execution methods
required to insert a job. Additionally, when checking for a unique job,
the driver always opens a transaction with `self.conn.begin_nested()`,
which will error unless in a transaction already.

In the Ruby client there's only a single `River#insert` method because
convention in that language is to have a lot of global stuff going on,
and the current transaction is added to local thread storage by Sequel
or ActiveRecord so that we can determine whether or not we're already in
a transaction from within the client.

SQLAlchemy is a little different. After opening a transaction, you're
expected to track the session object for perform operations in it:

    with session.begin():
        session.add(some_object())
        session.add(some_other_object())

So here I'm proposing a client API that looks a little more like the Go
API, with a pair of insert functions for `#insert` and `#insert_tx`, the
former doing an insert on the connection originally passed to driver,
and the latter doing an insert on a transaction sent as argument.

So this style of top-level insertion is now possible:

    engine = sqlalchemy.create_engine(database_url)
    client = riverqueue.Client(riversqlalchemy.Driver(engine))

    insert_res = client.insert(
        MyJobArgs(),
        insert_opts=riverqueue.InsertOpts(
            unique_opts=riverqueue.UniqueOpts(by_period=900)
        ),
    )

The client opens a new transaction on the driver's engine, and inserts
the job there.

Insertion on a particular transaction continues to be possible, and now
without having to reinitialize the client/driver:

    with engine.begin() as session:
        insert_res = client.insert_tx(
            session,
            MyJobArgs(),
            insert_opts=riverqueue.InsertOpts(
                unique_opts=riverqueue.UniqueOpts(by_period=900)
            ),
        )
        print(insert_res)