Async client for use with Python's asynchronous I/O + examples
Here we add support for Python's built-in asyncio [1], a feature that's
commonly requested for library APIs because it enables very performant
code built on asynchronous I/O.

The change creates a second client class called `AsyncClient` along with
an `riversqlalchemy.AsyncDriver` to go with the original driver. This
design is based on SQLAlchemy's and the code that the Python sqlc plugin
generates. The APIs for the pair of clients and drivers are identical,
but one set has `async` annotations on everything and takes appropriate
inputs (like an async SQLAlchemy engine instead of synchronous one).

The final usage looks pretty close to non-async code (which I think is a
pretty reasonable outcome overall):

    engine = sqlalchemy.ext.asyncio.create_async_engine(dev_database_url(is_async=True))
    client = riverqueue.AsyncClient(riversqlalchemy.AsyncDriver(engine))

    insert_res = await client.insert(
        SortArgs(strings=["whale", "tiger", "bear"]),
        insert_opts=riverqueue.InsertOpts(
            unique_opts=riverqueue.UniqueOpts(by_period=900)
        ),
    )
    print(insert_res)

I also added a new `examples/` package that shows code for how to do
various things with River Python, and makes sure they're run in CI and
get static analysis from MyPy.

[1] https://docs.python.org/3/library/asyncio.html