Drop Docker-based tests in favor of normal Postgres
Drop Docker-based tests (or at least within the Python program) in favor
of using normal Postgres like the other River projects are doing. I'm
having a lot of trouble getting Docker working in conjunction with
async. It's probably possible to get it working, but the testcontainers
docs are subpar, and I'm having to iterate failing tests in GitHub
Actions which is awful.