Separate GitHub Actions job for format, lint, and type check
Move format, lint, and type check to a separate GitHub Actions job. This
is partly because some of these steps are a little slow so it's better
if they're running parallel, but also it's annoying to lose a whole test
run when something dumb like a lint problem fails the job early. Now, we
still see the lint failure, but get to see test failures in case we have
any of those.