Add usage instructions README + job state constants + change insert many return
Flesh out the README to add detailed usage instructions. The README also
appears on PyPI, so it's nice to have extra information on what to do
there. It'll also be very closely mirrored on River's docs site.

Also, add job state constants like `JOB_STATE_AVAILABLE`, similar to
what's available in Ruby in case anyone needs them (they're useful for
use in `UniqueOpts`).

Lastly, change the return value of `insert_many`/`insert_many_tx` to be
the number of rows inserted instead of a list of inserted jobs to match
what's returned in Go and Ruby. The rationale here is that if someone is
using the bulk insert functions then they're probably concerned about
performance, so it's probably not worth the cost of returning the whole
set of jobs back, which are probably used infrequently anyway. These
functions aren't implemented yet anyway, so the change isn't a problem.