Implement insert many functions (#22)
Here, finish the implementations on `#insert_many` and `#insert_many_tx`,
which were previously marked as not yet finished. The Python plugin for
sqlc doesn't implement `copyfrom`, so we instead fall back on an
alternative approach from an experimental `database/sql` branch I have,
which implements an insert many operation using a combination of Postgres
arrays and `unnest`.

We remove some `Optional` annotations on `JobInsertParams` for properties
that are always set by the client. (Necessary to make MyPy's checks pass
properly.)