Use real enum for `JobState` instead of many constant
When I put in job state constants in #19, I didn't think to check first
whether there was a concept of real enums in Python. Turns out there is,
and we should use one instead of a whole bunch of raw constants.

This is a breaking change, but the job states have existed for so little
time that I doubt it'll break anybody, so it's worth doing it in this
instance.

And a few other small changes:

* Add typing to all exported constants.

* Fix an "invalid escape sequence" in the tag validation regex stemming
  from the `\A` in it by making the string a "raw string" with `r`.