Group async and sync tests into separate test classes (#29)
Group async and sync tests into separate test classes like `TestClient`
and `TestAsyncClient`. Along with organizing things more clearly, the
major advantage of this is that we can localize fixtures to the specific
classes where they're useful. So instead of `client` and `client_async`,
we just have `client` in both both test classes. This is not only more
clear, but greatly reduces errors because it's not possible to use the
wrong fixture, and makes tests more copy/pastable.