Use test-only sqlc queries instead of driver queries in test cases
Introduce a `JobGetByID` in sqlc which we switch to using in tests
instead of using the inelegant `job_get_by_kind_and_unique_properties()`
from the driver. This also lets us remove `driver` as a test fixture,
thus simplifying the whole testing set up and making it easier to
understand.