Fast path for unique insertion
Updates the Python client to be compatible with the fast unique
insertion added to the main River in [1] which uses a unique index
instead of advisory lock + fetch as long as uniqueness is constrained to
the default set of unique job states.

Also, not so much by design originally, but upgrade to sqlc v1.27.0 as
we have going in other River projects.

[1] https://github.com/riverqueue/river/pull/451